Mensajes confirmacion

+++ src/pages/panel/ChatmessagesPanel.tsx
@@ -6,6 +6,8 @@ import { toast } from "react-toastify";
 import { ChatMessage, PaginatedResponse } from "../../types/GeneralTypes";
 import { Link, useParams } from "react-router";
 import { ChatroomsApi } from "../../features/chatrooms/chatroomsApi";
+import { confirmAlert } from 'react-confirm-alert'; 
+import 'react-confirm-alert/src/react-confirm-alert.css';
 
 function ChatMessagesPanel() {
   const { chatroomId } = useParams<{ chatroomId: string }>();
@@ -61,15 +63,31 @@ function ChatMessagesPanel() {
   };
 
   const handleDelete = (messageId: string) => {
-    ChatroomsApi.deleteMessage(messageId)
-      .then(() => {
-        setMessages(messages.filter(m => m.message.id !== messageId));
-        toast.success("Message deleted");
-      })
-      .catch(error => {
-        toast.error("Failed to delete message");
-        console.error(error);
-      });
+    
+    confirmAlert({
+          title: 'Delete Message',
+          message: 'This action cannot be undone',
+          buttons: [
+            {
+              label: 'Yes',
+              onClick: () => {
+                  ChatroomsApi.deleteMessage(messageId)
+                    .then(() => {
+                      setMessages(messages.filter(m => m.message.id !== messageId));
+                      toast.success("Message deleted");
+                    })
+                    .catch(error => {
+                      toast.error("Failed to delete message");
+                      console.error(error);
+                    });
+                  }
+                },
+                {
+                  label: 'No',
+                    onClick: () => {}
+                }
+              ]
+          });
   };
 
   return (
